Refactor renderers and header of TabularSheet
