+++ src/components/TabularSheet/styles.scss
@@ -3,41 +3,8 @@
 .tabular-sheet {
     background-color: $color-foreground;
 
-    .header {
-        position: relative;
-        width: 100%;
-        height: 100%;
-        user-select: none;
-
-        .title {
-            padding: $spacing-small;
-        }
-
-        .edit {
-            display: none;
-            position: absolute;
-            top: 0;
-            right: -10px;
-        }
-
-        .loading-container {
-            position: relative;
-            margin-left: $spacing-small;
-        }
-
-        &:hover {
-            .edit {
-                display: initial;
-            }
-        }
-    }
-
     .cell {
         display: block;
         padding: $spacing-small;
-
-        &.number {
-            text-align: right;
-        }
     }
 }
